+++ NEWS.md
@@ -1,4 +1,4 @@
-# mlr3tuning (development version)
+# mlr3tuning 1.1.0
 
 * fix: The `as_data_table()` functions do not unnest the `x_domain` colum anymore by default.
 * fix: `to_tune(internal = TRUE)` now also works if non-internal tuning parameters require have an `.extra_trafo`.